+++ multitld/src/core/Box.cc
@@ -44,7 +44,7 @@ bool Box::isStandard() {
     return sum == roundedSum;
 }
 
-Box* Box::move(float dx, float dy) {
+Box* Box::move(double dx, double dy) {
     Box* box = new Box(id, x1 + dx, y1 + dy, x2 + dx, y2 + dy);
     return box;
 }
